+++ server/src/test/java/org/opensearch/action/admin/indices/tiering/TransportHotToWarmTieringActionTests.java
@@ -56,7 +56,7 @@ public class TransportHotToWarmTieringActionTests extends OpenSearchIntegTestCas
     @Before
     public void setup() {
         internalCluster().startClusterManagerOnlyNode();
-        internalCluster().ensureAtLeastNumSearchAndDataNodes(1);
+        internalCluster().ensureAtLeastNumWarmAndDataNodes(1);
         long bytes = new ByteSizeValue(1000, ByteSizeUnit.KB).getBytes();
         final MockInternalClusterInfoService clusterInfoService = getMockInternalClusterInfoService();
         clusterInfoService.setDiskUsageFunctionAndRefresh((discoveryNode, fsInfoPath) -> setDiskUsage(fsInfoPath, bytes, bytes - 1));